fix relationship between order and payment, shipment

+++ src/main/java/com/javatechie/repository/ItemViewedRepository.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 
 @Repository
 public interface ItemViewedRepository extends JpaRepository<ItemViewedEntity, Long> {
-    Optional<ItemViewedEntity> findByItemIdAndUserId(Integer itemId, Integer userId);
+    Optional<ItemViewedEntity> findByItemIdAndUserId(Long itemId, Integer userId);
     @Query(value = "select * from item_view where (:favorite is null or favorite = :favorite) and (:viewed is null or viewed = :viewed) and user_id = :userId", nativeQuery = true)
     List<ItemViewedEntity> findAllByFavoriteOrViewedAndUser(@Param("favorite")Integer favorite, @Param("viewed")Integer viewed, @Param("userId")Integer userId);
 }
